backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
statsmodels
SQLAlchemy
python-multipart
pydantic
PyJWT
joblib
scipy
pyarrow
openpyxl
xlrd
passlib[bcrypt]
python-dotenv
email-validator
